fixed sell cuz ima dummy

+++ src/GameCoreInterface.java
@@ -130,7 +130,7 @@ public interface GameCoreInterface {
      * @param shopId The ID of the shop the player is selling an item to
      * @param item The item the player is selling (eventually will be an Item obj)
      */
-    public int sellItem(String name, int shopId, String item);
+    public float sellItem(String name, int shopId, String item);
     
     /**
      * Returns a Shop's inventory as a formatted string
