+++ src/GameObject.java
@@ -285,7 +285,7 @@ public class GameObject extends UnicastRemoteObject implements GameObjectInterfa
      * @param shopId The ID of the shop the player is selling an item to
      * @param item The item the player is selling (eventually will be an Item obj)
      */
-    public int sellItem(String name, int shopId, String item) throws RemoteException{
+    public float sellItem(String name, int shopId, String item) throws RemoteException{
     	return core.sellItem(name, shopId, item);
     }
