+++ src/ShopClient.java
@@ -133,7 +133,7 @@ public class ShopClient {
 	
 	//In terms of the player selling items
 	public void sell(String item) throws RemoteException{
-		int val = remote.sellItem(this.player, this.id, item);
+		float val = remote.sellItem(this.player, this.id, item);
 		if (val != 0) {
 			System.out.println("You have sold " + item + " for $" + val + ".00.");
 		}
